+++ Game/Map/TileLocator.cs
@@ -106,7 +106,7 @@ namespace Game.Map
             } while (!doSelf && (x == ox && y == oy));
         }
 
-        public virtual void ForeachObject(uint ox, uint oy, byte radius, bool doSelf, DoWork work, object custom)
+        public virtual void ForeachObject(uint ox, uint oy, int radius, bool doSelf, DoWork work, object custom)
         {
             byte mode = (byte)(oy%2 == 0 ? 0 : 1);
